requirements.txt
h5py==3.15.1
joblib==1.5.3
matplotlib==3.10.8
natsort==8.4.0
numpy==2.4.2
opencv-python==4.13.0.92
pandas==3.0.0
pillow==12.1.0
PyYAML==6.0.3
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
tqdm==4.67.3
